requirements.txt
rethinkdb
python-oauth2
flask
